+++ index.js
@@ -1,4 +1,0 @@
-module.exports = (bundler) => {
-  // Register ourselves to handle .njk files
-  bundler.addAssetType('njk', require.resolve('./src/NunjucksAsset'));
-};