+++ test/tests/option-asset-type-dynamic/nunjucks.config.js
@@ -1,0 +1,8 @@
+'use strict'
+
+module.exports = {
+    data({ filename }) { return { name: filename } },
+    assetType({ baseExt, dirname }) {
+        return baseExt || dirname
+    }
+}